Add initial version of ImageBuilder using Tekton Pipelines
