+++ docs/mico-core/java/io/github/ust/mico/core/service/imagebuilder/BuildList.rst
@@ -1,6 +1,6 @@
 .. java:import:: io.fabric8.kubernetes.client CustomResourceList
 
-.. java:import:: io.github.ust.mico.core.service.imagebuilder.buildtypes Build
+.. java:import:: io.github.ust.mico.core.service.imagebuilder.knativebuild.buildtypes Build
 
 BuildList
 =========
