+++ docs/mico-core/java/io/github/ust/mico/core/service/imagebuilder/buildtypes/ArgumentSpec.rst
@@ -11,7 +11,7 @@
 ArgumentSpec
 ============
 
-.. java:package:: io.github.ust.mico.core.service.imagebuilder.buildtypes
+.. java:package:: io.github.ust.mico.core.service.imagebuilder.knativebuild.buildtypes
    :noindex:
 
 .. java:type:: @JsonInclude @Data @NoArgsConstructor @AllArgsConstructor @Accessors public class ArgumentSpec
