+++ docs/mico-core/java/io/github/ust/mico/core/service/imagebuilder/buildtypes/BuildStep.rst
@@ -19,7 +19,7 @@
 BuildStep
 =========
 
-.. java:package:: io.github.ust.mico.core.service.imagebuilder.buildtypes
+.. java:package:: io.github.ust.mico.core.service.imagebuilder.knativebuild.buildtypes
    :noindex:
 
 .. java:type:: @JsonInclude @JsonIgnoreProperties @Data @NoArgsConstructor @AllArgsConstructor @Accessors public class BuildStep
