+++ install/kubernetes/mico-build-bot.yaml
@@ -1,13 +1,13 @@
 ---
 apiVersion: v1
 kind: Secret
-type: kubernetes.io/basic-auth
 metadata:
   name: build-bot-dockerhub-secret
   namespace: mico-build-bot
   annotations:
-    build.knative.dev/docker-0: https://index.docker.io/v1/
-data:
+    tekton.dev/docker-0: https://index.docker.io/v1/
+type: kubernetes.io/basic-auth
+stringData:
   username: ${DOCKERHUB_USERNAME_BASE64}
   password: ${DOCKERHUB_PASSWORD_BASE64}
 ---
