+++ mico-core/src/main/java/io/github/ust/mico/core/service/ImageBuilderHealthIndicator.java
@@ -20,6 +20,7 @@
 package io.github.ust.mico.core.service;
 
 import io.github.ust.mico.core.service.imagebuilder.ImageBuilder;
+import io.github.ust.mico.core.service.imagebuilder.knativebuild.KnativeBuildController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.HealthIndicator;
@@ -31,7 +32,7 @@ public class ImageBuilderHealthIndicator implements HealthIndicator {
     private final ImageBuilder imageBuilder;
 
     @Autowired
-    public ImageBuilderHealthIndicator(ImageBuilder imageBuilder) {
+    public ImageBuilderHealthIndicator(KnativeBuildController imageBuilder) {
         this.imageBuilder = imageBuilder;
     }
