--- mico-core/src/main/java/io/github/ust/mico/core/service/imagebuilder/buildtypes/Volume.java
@@ -1,51 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.github.ust.mico.core.service.imagebuilder.buildtypes;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-/**
- * Volume represents a named volume in a pod that may be accessed by any container in the pod.
- */
-@JsonInclude(JsonInclude.Include.NON_EMPTY)
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Accessors(chain = true)
-public class Volume {
-
-    /**
-     * Volume's name.
-     * Must be a DNS_LABEL and unique within the pod.
-     * More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
-     */
-    private String name;
-
-    /**
-     * VolumeSource represents the location and type of the mounted volume.
-     * If not specified, the Volume is implied to be an EmptyDir.
-     * This implied behavior is deprecated and will be removed in a future version.
-     */
-    private VolumeSource volumeSource;
-}
