+++ mico-core/src/main/java/io/github/ust/mico/core/service/imagebuilder/knativebuild/buildtypes/ArgumentSpec.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.github.ust.mico.core.service.imagebuilder.buildtypes;
+package io.github.ust.mico.core.service.imagebuilder.knativebuild.buildtypes;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
