+++ mico-core/src/main/java/io/github/ust/mico/core/service/imagebuilder/knativebuild/buildtypes/Volume.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.github.ust.mico.core.service.imagebuilder.buildtypes;
+package io.github.ust.mico.core.service.imagebuilder.knativebuild.buildtypes;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
@@ -26,17 +26,26 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
 /**
- * SourceSpec defines the input to the Build
+ * Volume represents a named volume in a pod that may be accessed by any container in the pod.
  */
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Accessors(chain = true)
-public class SourceSpec {
+public class Volume {
 
     /**
-     * Optional. Git represents source in a Git repository.
+     * Volume's name.
+     * Must be a DNS_LABEL and unique within the pod.
+     * More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
      */
-    private GitSourceSpec git;
+    private String name;
+
+    /**
+     * VolumeSource represents the location and type of the mounted volume.
+     * If not specified, the Volume is implied to be an EmptyDir.
+     * This implied behavior is deprecated and will be removed in a future version.
+     */
+    private VolumeSource volumeSource;
 }
