+++ mico-core/src/main/resources/application-dev.properties
@@ -32,11 +32,11 @@ spring.cloud.kubernetes.reload.enabled=false
 kubernetes.namespace-mico-workspace=mico-testing
 kubernetes.namespace-openfaas-workspace=openfaas
 kubernetes.build-bot.namespace-build-execution=mico-testing
-kubernetes.build-bot.docker-image-repository-url=docker.io/ustmico
+kubernetes.build-bot.docker-image-repository-url=docker.io/<<YOUR_ACCOUNT_NAME>>
 kubernetes.build-bot.docker-registry-service-account-name=build-bot-dockerhub
-kubernetes.build-bot.kaniko-executor-image-url=gcr.io/kaniko-project/executor
+kubernetes.build-bot.kaniko-executor-image-url=gcr.io/kaniko-project/executor@sha256:f87c11770a4d3ed33436508d206c584812cd656e6ed08eda1cff5c1ee44f5870
 kubernetes.build-bot.build-timeout=600
-kubernetes.build-bot.build-clean-up-by-undeploy=false
+kubernetes.build-bot.build-clean-up-by-undeploy=true
 
 # Prometheus
 kubernetes.prometheus.uri=http://localhost:9090/api/v1/query
