+++ mico-core/src/main/resources/application.properties
@@ -19,7 +19,7 @@
 
 # Application
 spring.application.name=mico-core
-server.port=8080
+server.port=5000
 
 # Profiles:
 # - 'dev':          typical development environment: with Kubernetes connection, real local Neo4j database
@@ -40,7 +40,7 @@ logging.level.springfox.documentation=WARN
 logging.level.io.lettuce.core=INFO
 
 # Cors Origins, It is a comma separated list
-cors-policy.allowed-origins=http://localhost:4200,http://localhost:8080
+cors-policy.allowed-origins=http://localhost:4200,http://localhost:8080,http://localhost:5000
 cors-policy.additional-allowed-methods=OPTIONS,PATCH
 
 # Kafka FaaS Connector
