+++ mico-core/src/test/java/io/github/ust/mico/core/MicoKubernetesClientTests.java
@@ -19,25 +19,57 @@
 
 package io.github.ust.mico.core;
 
+import java.net.PasswordAuthentication;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
-import io.fabric8.kubernetes.api.model.*;
+import io.fabric8.kubernetes.api.model.EnvVar;
+import io.fabric8.kubernetes.api.model.NamespaceBuilder;
+import io.fabric8.kubernetes.api.model.Pod;
+import io.fabric8.kubernetes.api.model.PodBuilder;
+import io.fabric8.kubernetes.api.model.SecretBuilder;
+import io.fabric8.kubernetes.api.model.Service;
+import io.fabric8.kubernetes.api.model.ServiceBuilder;
+import io.fabric8.kubernetes.api.model.ServiceList;
+import io.fabric8.kubernetes.api.model.ServicePort;
+import io.fabric8.kubernetes.api.model.ServiceStatusBuilder;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.DeploymentBuilder;
 import io.fabric8.kubernetes.api.model.apps.DeploymentList;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
-import io.github.ust.mico.core.TestConstants.*;
+import io.github.ust.mico.core.TestConstants.IntegrationTest;
 import io.github.ust.mico.core.broker.BackgroundJobBroker;
 import io.github.ust.mico.core.configuration.MicoKubernetesBuildBotConfig;
 import io.github.ust.mico.core.configuration.MicoKubernetesConfig;
 import io.github.ust.mico.core.exception.KubernetesResourceException;
 import io.github.ust.mico.core.exception.MicoApplicationNotFoundException;
-import io.github.ust.mico.core.model.*;
+import io.github.ust.mico.core.model.KubernetesDeploymentInfo;
+import io.github.ust.mico.core.model.MicoApplication;
+import io.github.ust.mico.core.model.MicoApplicationDeploymentStatus;
+import io.github.ust.mico.core.model.MicoApplicationJobStatus;
+import io.github.ust.mico.core.model.MicoEnvironmentVariable;
+import io.github.ust.mico.core.model.MicoInterfaceConnection;
+import io.github.ust.mico.core.model.MicoLabel;
+import io.github.ust.mico.core.model.MicoService;
 import io.github.ust.mico.core.model.MicoServiceBackgroundJob.Status;
+import io.github.ust.mico.core.model.MicoServiceDeploymentInfo;
+import io.github.ust.mico.core.model.MicoServiceInterface;
+import io.github.ust.mico.core.model.MicoServicePort;
+import io.github.ust.mico.core.model.MicoTopic;
+import io.github.ust.mico.core.model.MicoTopicRole;
+import io.github.ust.mico.core.model.OpenFaaSFunction;
 import io.github.ust.mico.core.persistence.KubernetesDeploymentInfoRepository;
 import io.github.ust.mico.core.persistence.MicoApplicationRepository;
 import io.github.ust.mico.core.persistence.MicoServiceDeploymentInfoRepository;
 import io.github.ust.mico.core.service.MicoKubernetesClient;
-import io.github.ust.mico.core.service.imagebuilder.ImageBuilder;
+import io.github.ust.mico.core.service.imagebuilder.knativebuild.KnativeBuildController;
 import io.github.ust.mico.core.util.CollectionUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
@@ -49,21 +81,39 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.net.PasswordAuthentication;
-import java.util.*;
-import java.util.stream.Collectors;
-
 import static com.spotify.hamcrest.optional.OptionalMatchers.emptyOptional;
 import static com.spotify.hamcrest.optional.OptionalMatchers.optionalWithValue;
-import static io.github.ust.mico.core.TestConstants.*;
-import static io.github.ust.mico.core.service.MicoKubernetesClient.*;
+import static io.github.ust.mico.core.TestConstants.ID_1;
+import static io.github.ust.mico.core.TestConstants.INSTANCE_ID;
+import static io.github.ust.mico.core.TestConstants.INSTANCE_ID_1;
+import static io.github.ust.mico.core.TestConstants.INSTANCE_ID_2;
+import static io.github.ust.mico.core.TestConstants.NAME;
+import static io.github.ust.mico.core.TestConstants.NAME_1;
+import static io.github.ust.mico.core.TestConstants.NAME_2;
+import static io.github.ust.mico.core.TestConstants.SERVICE_INTERFACE_NAME;
+import static io.github.ust.mico.core.TestConstants.SERVICE_INTERFACE_NAME_1;
+import static io.github.ust.mico.core.TestConstants.SERVICE_SHORT_NAME;
+import static io.github.ust.mico.core.TestConstants.SERVICE_SHORT_NAME_1;
+import static io.github.ust.mico.core.TestConstants.SERVICE_VERSION;
+import static io.github.ust.mico.core.TestConstants.SHORT_NAME;
+import static io.github.ust.mico.core.TestConstants.SHORT_NAME_1;
+import static io.github.ust.mico.core.TestConstants.SHORT_NAME_2;
+import static io.github.ust.mico.core.TestConstants.VERSION;
+import static io.github.ust.mico.core.service.MicoKubernetesClient.OPEN_FAAS_SECRET_DATA_PASSWORD_NAME;
+import static io.github.ust.mico.core.service.MicoKubernetesClient.OPEN_FAAS_SECRET_DATA_USERNAME_NAME;
+import static io.github.ust.mico.core.service.MicoKubernetesClient.OPEN_FAAS_SECRET_NAME;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.hamcrest.core.IsCollectionContaining.hasItem;
 import static org.hamcrest.core.IsCollectionContaining.hasItems;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.BDDMockito.any;
 import static org.mockito.BDDMockito.given;
 
@@ -78,35 +128,25 @@ public class MicoKubernetesClientTests {
     private static final String LABEL_VERSION_KEY = LABEL_PREFIX + "version";
     private static final String LABEL_INTERFACE_KEY = LABEL_PREFIX + "interface";
     private static final String LABEL_INSTANCE_KEY = LABEL_PREFIX + "instance";
-
-    @Rule
-    public KubernetesServer mockServer = new KubernetesServer(false, true);
-
-    @MockBean
-    private MicoKubernetesConfig micoKubernetesConfig;
-
-    @MockBean
-    private MicoKubernetesBuildBotConfig micoKubernetesBuildBotConfig;
-
-    @MockBean
-    private ImageBuilder imageBuilder;
-
-    @MockBean
-    private BackgroundJobBroker backgroundJobBroker;
-
-    @MockBean
-    private MicoApplicationRepository applicationRepository;
-
-    @MockBean
-    private MicoServiceDeploymentInfoRepository serviceDeploymentInfoRepository;
-
-    @MockBean
-    private KubernetesDeploymentInfoRepository kubernetesDeploymentInfoRepository;
-
-    private MicoKubernetesClient micoKubernetesClient;
-
     private static final String testNamespace = "test-namespace";
     private static final String buildTestNamespace = "test-namespace";
+    @Rule
+    public KubernetesServer mockServer = new KubernetesServer(false, true);
+    @MockBean
+    private MicoKubernetesConfig micoKubernetesConfig;
+    @MockBean
+    private MicoKubernetesBuildBotConfig micoKubernetesBuildBotConfig;
+    @MockBean
+    private KnativeBuildController imageBuilder;
+    @MockBean
+    private BackgroundJobBroker backgroundJobBroker;
+    @MockBean
+    private MicoApplicationRepository applicationRepository;
+    @MockBean
+    private MicoServiceDeploymentInfoRepository serviceDeploymentInfoRepository;
+    @MockBean
+    private KubernetesDeploymentInfoRepository kubernetesDeploymentInfoRepository;
+    private MicoKubernetesClient micoKubernetesClient;
 
     @Before
     public void setUp() {
@@ -561,7 +601,7 @@ public class MicoKubernetesClientTests {
             .create(new PodBuilder()
                 .withNewMetadata()
                 .withLabels(CollectionUtils.mapOf(
-                    ImageBuilder.BUILD_CRD_GROUP + "/buildName",
+                    KnativeBuildController.BUILD_CRD_GROUP + "/buildName",
                     imageBuilder.createBuildName(micoService)))
                 .endMetadata()
                 .build());
@@ -580,7 +620,7 @@ public class MicoKubernetesClientTests {
         List<Pod> actualPods = mockServer.getClient()
             .pods()
             .inAnyNamespace()
-            .withLabel(ImageBuilder.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
+            .withLabel(KnativeBuildController.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
             .list().getItems();
         assertNull("Expected Kubernetes deployment is deleted", actualDeployment);
         assertNull("Expected Kubernetes service is deleted", actualService);
@@ -665,7 +705,7 @@ public class MicoKubernetesClientTests {
             .create(new PodBuilder()
                 .withNewMetadata()
                 .withLabels(CollectionUtils.mapOf(
-                    ImageBuilder.BUILD_CRD_GROUP + "/buildName",
+                    KnativeBuildController.BUILD_CRD_GROUP + "/buildName",
                     imageBuilder.createBuildName(micoService)))
                 .endMetadata()
                 .build());
@@ -684,7 +724,7 @@ public class MicoKubernetesClientTests {
         List<Pod> actualPods = mockServer.getClient()
             .pods()
             .inAnyNamespace()
-            .withLabel(ImageBuilder.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
+            .withLabel(KnativeBuildController.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
             .list().getItems();
         // It's not possible to check if the actual replicas are valid, because they are not applied immediately.
         // Therefore just test if the Kubernetes resources still exist and there are no error during the scale in.
@@ -767,7 +807,7 @@ public class MicoKubernetesClientTests {
             .create(new PodBuilder()
                 .withNewMetadata()
                 .withLabels(CollectionUtils.mapOf(
-                    ImageBuilder.BUILD_CRD_GROUP + "/buildName",
+                    KnativeBuildController.BUILD_CRD_GROUP + "/buildName",
                     imageBuilder.createBuildName(micoService)))
                 .endMetadata()
                 .build());
@@ -786,7 +826,7 @@ public class MicoKubernetesClientTests {
         List<Pod> actualPods = mockServer.getClient()
             .pods()
             .inAnyNamespace()
-            .withLabel(ImageBuilder.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
+            .withLabel(KnativeBuildController.BUILD_CRD_GROUP + "/buildName", imageBuilder.createBuildName(micoService))
             .list().getItems();
         // It's not possible to check if the actual replicas are valid, because they are not applied immediately.
         // Therefore just test if the Kubernetes resources still exist and there are no error during the scale in.
@@ -869,7 +909,6 @@ public class MicoKubernetesClientTests {
             "apiVersion: \"apps/v1\"\n" +
             "kind: \"Deployment\"\n" +
             "metadata:\n" +
-            "  annotations: {}\n" +
             "  labels:\n" +
             "    ust.mico/instance: \"" + instanceId + "\"\n" +
             "    ust.mico/name: \"" + SERVICE_SHORT_NAME + "\"\n" +
@@ -882,7 +921,6 @@ public class MicoKubernetesClientTests {
             "apiVersion: \"v1\"\n" +
             "kind: \"Service\"\n" +
             "metadata:\n" +
-            "  annotations: {}\n" +
             "  labels:\n" +
             "    ust.mico/interface: \"" + SERVICE_INTERFACE_NAME + "\"\n" +
             "    ust.mico/instance: \"" + instanceId + "\"\n" +
@@ -890,10 +928,8 @@ public class MicoKubernetesClientTests {
             "    ust.mico/version: \"" + SERVICE_VERSION + "\"\n" +
             "  name: \"" + micoKubernetesClient.createServiceName(micoServiceDeploymentInfo, micoServiceInterface) + "\"\n" +
             "  namespace: \"" + testNamespace + "\"\n" +
-            "spec:\n" +
-            "  selector: {}\n", actualYaml);
+            "spec: {}\n", actualYaml);
     }
-
 
     @Test
     public void testGetOpenFaaSCredentialsTest() {
@@ -931,7 +967,6 @@ public class MicoKubernetesClientTests {
         String testServiceName = "testservice";
         micoKubernetesClient.getPublicIpOfKubernetesService(testServiceName, testNamespace);
     }
-
 
     @Test
     public void getExternalIpOfService() throws Exception {
@@ -991,9 +1026,11 @@ public class MicoKubernetesClientTests {
             getServiceBuilderWithNameAndNamespace(testServiceName)
                 .withNewSpec()
                 .addNewPort()
-                .withNewPort(port)
+                .withPort(port)
                 .endPort()
-                .endSpec().build());
+                .endSpec()
+                .build()
+        );
         List<Integer> ports = micoKubernetesClient.getPublicPortsOfKubernetesService(testServiceName, testNamespace);
         assertThat(ports, hasItem(port));
         assertThat(ports, hasSize(1));
@@ -1008,12 +1045,14 @@ public class MicoKubernetesClientTests {
             getServiceBuilderWithNameAndNamespace(testServiceName)
                 .withNewSpec()
                 .addNewPort()
-                .withNewPort(port)
+                .withPort(port)
                 .endPort()
                 .addNewPort()
-                .withNewPort(port2)
+                .withPort(port2)
                 .endPort()
-                .endSpec().build());
+                .endSpec()
+                .build()
+        );
         List<Integer> ports = micoKubernetesClient.getPublicPortsOfKubernetesService(testServiceName, testNamespace);
         assertThat(ports, hasItems(port, port2));
         assertThat(ports, hasSize(2));
